Unify pasador frame analysis with beam simulator
